+++ protocols/util/dependency_manager.py
@@ -4,11 +4,8 @@ import os.path
 import inspect
 from protocols.util.singleton import Singleton
 
-VERSION_PARTICIPANT_103_TEST = "1.0.3-participant-tests"
-VERSION_430 = "4.3.0-SNAPSHOT"
-VERSION_410 = "4.1.0"
+VERSION_500 = "5.0.0"
 VERSION_400 = "4.0.0"
-VERSION_310 = "3.1.0"
 VERSION_300 = "3.0.0"
 VERSION_210 = "2.1.0"